chore: make fmt

+++ src/test/unit/KeyRegistrarUnit.t.sol
@@ -398,7 +398,7 @@ contract KeyRegistrarUnitTests_registerKey_BN254 is KeyRegistrarUnitTests {
         vm.prank(avs1);
         keyRegistrar.configureOperatorSet(operatorSet, CurveType.BN254);
 
-        uint256 len = vm.randomUint(0, 256);
+        uint len = vm.randomUint(0, 256);
         vm.assume(len != 192);
         bytes memory keyData = new bytes(len);
         bytes memory signature = _generateBN254Signature(operator1, operatorSet, bn254Key1, bn254PrivKey1);
@@ -406,7 +406,7 @@ contract KeyRegistrarUnitTests_registerKey_BN254 is KeyRegistrarUnitTests {
         vm.prank(operator1);
         vm.expectRevert(InvalidKeyFormat.selector);
         keyRegistrar.registerKey(operator1, operatorSet, keyData, signature);
-    } 
+    }
 
     function test_revert_invalidSignatureLength() public {
         OperatorSet memory operatorSet = _createOperatorSet(avs1, DEFAULT_OPERATOR_SET_ID);
@@ -414,7 +414,7 @@ contract KeyRegistrarUnitTests_registerKey_BN254 is KeyRegistrarUnitTests {
         vm.prank(avs1);
         keyRegistrar.configureOperatorSet(operatorSet, CurveType.BN254);
 
-        uint256 len = vm.randomUint(0, 256);
+        uint len = vm.randomUint(0, 256);
         vm.assume(len != 64);
         bytes memory malformedSignature = new bytes(len);
